+++ Simple Game Library/CHANGES.md
@@ -6,6 +6,7 @@ Date: *2016-04-??*
 - Added more JavaDoc comments.
 - Added constructor to `Game` class taking an initial game state as parameter.
 - Fixed bug in Java2dApplication causing the `resize` method not to be called on startup.
+- Fixed bug causing the application to stop working when toggling full screen mode.
 
 ## Version 1.2.0
 Date: *2016-04-18*