+++ Simple Game Library/CHANGES.md
@@ -7,6 +7,9 @@ Date: *2016-04-??*
 - Added constructor to `Game` class taking an initial game state as parameter.
 - Fixed bug in Java2dApplication causing the `resize` method not to be called on startup.
 - Fixed bug causing the application to stop working when toggling full screen mode.
+- Add constants to math utilities used to convert nanoseconds.
+- Added `getAspectRatio` method to `Graphics` interface.
+- Added some additional JavaDoc comments.
 
 ## Version 1.2.0
 Date: *2016-04-18*