+++ Simple Game Library/CHANGES.md
@@ -13,7 +13,7 @@ Date: *2016-04-??*
 - Added `getAspectRatio` method to `Graphics` interface.
 - Added some additional JavaDoc comments.
 - Improved Game class, `switchState` has now a parameter if the state's  `resize` method should be called.
-
+- Fixed bug in input system causing an IndexOutOfBoundsException.
 
 ## Version 1.2.0
 Date: *2016-04-18*